re-factor -X options
This change replaces redunant -X option code and adds addXdetails function to handle it. it reduces the number of lines and is a little cleaner.
code also includes extra details even if there is no description (replaces description with title for looks) and adds season and episode numbers to extra details.